Most basic aligner, can be improved by:  * precomputing the matching words (replaces n! comparisons by n^2 comparisons and n! lookups in the worst case) [1]  * different word matchers (for instance: we could take synonyms into account)  * different align() heuristic (currently: __maximal consecutive alignments__, could be maximal crossing edges)  * align punctuation
This two last things should be implemented with external objects (strategy pattern) to be able to instantiate them at runtime as an option.

Currently makes the longuest alignment line by line from two (w)fstreams.
The result is in std::vector<int> _final_alignment with 
_final_alignment[indice _word_right] = indice of the mapped _word_left 
when this indice is -1, it means that this words are unaligned, this gives freedom for the hypothesis generator.

/!\ this doesn't align all that can be aligned, see examples. Will make more alignments in a next version (with [1]).

To try:
cd apertium-combine/memt/
make
./aligner ../test/apertium.in ../test/moses.in