add a string directing to beta.apertium.org (#483)
**Description:**
This PR addresses issue #361 and also builds upon and supersedes the
changes proposed in PR #422.

**Changes Made:**
- Modified `config.ts` by adding `showMoreLanguagesLink` and specified
its type as `boolean` in `types.ts`. Also, added `more_languages` to
`stringReplacements`.
- Modified `AboutModal.tsx` by conditionally rendering `more_languages`
based on the value of `shouldDisplayString`, either true or false.
- In `index.test.tsx`, modified a test to check if, when the About
dialog is opened and `showMoreLanguagesLink` is set to `true`, it should
render the `more_languages`. Similarly, when `showMoreLanguagesLink` is
set to `false` and the About dialog is opened, it should not render the
`more_languages`.